+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # TODO 
 
-- [ ] fix bug meld version (shell sh)
 - [x] feat mode debug
 - [x] feat premier mot de version ou premiere ligne ? cf8 /etc/osrelease gradle
-- [ ] fix max one category
+- [x] fix max one category
 - [x] add app running
+- [x] limit to zsh or bash
+- [ ] test category and app exists